requirements.txt
appdirs==1.4.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==0.12.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymessenger==0.0.7.0
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
urllib3==1.26.4
virtualenv==20.4.4
Werkzeug==1.0.1
gunicorn==19.6.0
wit == 6.0.0
pytz==2019.1
beautifulsoup4==4.9.3
lxml==4.6.3
psycopg2==2.8.6
pyjokes==0.6.0
fbmessenger==6.0.0
rivescript==1.15.0
python-dateutil==2.8.1
tabulate==0.8.9
PyGithub==1.55